requirements.txt
crewai>=0.80.0
crewai-tools>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
